requirements.txt
pip==20.1.1
setuptools==47.1.0
absl-py==2.1.0
adal==1.2.7
applicationinsights==0.11.10
argcomplete==3.1.2
astunparse==1.6.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-graphrbac==0.61.1
azure-identity==1.15.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azureml-automl-core==1.55.0.post2
azureml-core==1.55.0.post2
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.2
azureml-dataset-runtime==1.55.0
azureml-pipeline==1.55.0
azureml-pipeline-core==1.55.0
azureml-pipeline-steps==1.55.0
azureml-sdk==1.55.0
azureml-telemetry==1.55.0
azureml-train-automl-client==1.55.0.post1
azureml-train-core==1.55.0
azureml-train-restclients-hyperdrive==1.55.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.3
cachetools==5.3.3
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
cloudpickle==2.2.1
contextlib2==21.6.0
cryptography==42.0.8
docker==6.1.3
flatbuffers==24.3.25
fusepy==3.0.1
gast==0.4.0
google-auth==2.31.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.8.0
humanfriendly==10.0
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
isodate==0.6.1
jeepney==0.8.0
jmespath==1.0.1
jsonpickle==3.2.2
jsonschema==4.17.3
keras==2.11.0
knack==0.11.0
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
paramiko==3.4.0
pathspec==0.11.2
pkginfo==1.10.0
pkgutil-resolve-name==1.3.10
portalocker==2.7.0
protobuf==3.19.6
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyreadline==2.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing-extensions==4.7.1
urllib3==2.0.7
websocket-client==1.6.1
Werkzeug==2.2.3
wheel==0.42.0
wrapt==1.16.0
zipp==3.15.0
